requirements.txt
# Credit Risk Analysis - Production Requirements
# Azure and Databricks deployment ready

# Core Framework
flask>=3.0.3
flask-cors>=4.0.1
flask-limiter>=3.7.0
plotly==5.17.0
pandas>=2.1.4
numpy>=1.25.0

# Data Processing  
openpyxl==3.1.2
xlsxwriter==3.1.9
rapidfuzz==3.6.0

# Azure Dependencies
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.19.0
azure-monitor-opentelemetry==1.2.0

# Essential packages for Azure deployment
Werkzeug>=3.0.1
Jinja2>=3.1.2
itsdangerous>=2.1.2
MarkupSafe>=2.1.3

# Databricks dependencies
databricks-sdk>=0.12.0
pyspark>=3.4.0
delta-spark>=2.4.0

# Machine Learning
scikit-learn>=1.3.0
mlflow>=2.7.0

# LangGraph and AI
langgraph>=0.0.50
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5

# HTTP and API
requests>=2.31.0
httpx>=0.25.2
urllib3>=2.1.0

# File Operations and Locking
portalocker>=2.7.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
streamlit-aggrid>=1.1.8

# Future Phase 2 Dependencies (commented out for now)
# langchain>=0.1.0
# langgraph>=0.1.0
# openai>=1.0.0
# beautifulsoup4>=4.12.0
# requests>=2.31.0
# chromadb>=0.4.0

# Future Phase 3 Dependencies (commented out for now)
# databricks-sdk>=0.12.0
# mlflow>=2.8.0
# delta-spark>=3.0.0

# NOTE: Removed duplicate legacy dependencies - using more recent versions above
